requirements.txt
aiogram==3.10.0
telethon==1.36.0
sqlalchemy==2.0.30
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
psycopg2-binary
matplotlib==3.8.4
google-generativeai==0.5.2